blk: fix allocator ordering issues (thanks mcf)
logappend inserts LogAlloc1 in case it needed to
allocate a new log block.

However, these new log blocks are allocated from
the root of the avl tree, and compresslog traverses
starting from avlmin; the current code inserts into
the current log segment, instead of the active one,
which means that logappend could insert a LogAlloc1
for the new log block before the freeing of the block
would be allocated.

Additionally, we chain the wrong block when compressing
the allocation log.

This fixes both issues.