fs: fix invalid Dent created with partial walk
If we walked some number of components, but not the full path, we
do not need the Dent. Calling getdent here leaks a reference, and
may potentially create an invalid Dent in memory.

d.name, backed by kvbuf, may have been overwritten during a btlookup
that encounters an Oinsert followed by Odelete. If there are no
other references to the Dent for the last successful wname, this
results in the creation and leak of a Dent with an incorrect pqid-name
pair, breaking subsequent walks to that wname.