blk: move from hacky qsbr to ebr
the quiescent state based reclamation was too home-grown
for my taste, complicated, and not at all lock free; moving
to epoch based reclamation simplifies the code, does a far
more textbook implementation of the algorithm, and removes
locks around the active generations.

this also serves as a testbed for the atomic APIs we want
to support.